ya ni4ego ne sdelal :)

+++ Assets/Scripts/Helpers.cs
@@ -1,5 +1,7 @@
+using System;
 using Fungus;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 public class Helpers : MonoBehaviour
 {
@@ -11,8 +13,38 @@ public class Helpers : MonoBehaviour
     {
         Application.OpenURL("https://vk.com/1rgt1"); 
     }
+
+    public void Continue()
+    {
+        FungusManager.Instance.SaveManager.Load("save_data");
+    }
+    
     public void DeleteHistory()
     {
-        SaveManager.Delete("save_data");
+        if (FungusManager.Instance.SaveManager.SaveDataExists("save_data"))
+        {
+            try
+            {
+                SaveManager.Delete("save_data");
+            }
+            catch (Exception)
+            {
+                // ignored
+            }
+        }
+        FungusManager.Instance.SaveManager.ClearHistory();
+        var _step = FungusManager.Instance.GlobalVariables.GetOrAddVariable("BlockStep", 0, typeof(IntegerVariable));
+        _step.Apply(SetOperator.Assign, 0);
     }
+
+    public void LoadAbout()
+    {
+        SceneManager.LoadScene("About");
+    }
+
+    public void LoadMenu()
+    {
+        SceneManager.LoadScene("MainMenu");
+    }
+    
 }
